+++ changelog.md
@@ -1,3 +1,7 @@
 ### General
 
 - Fixed improper rule categories
+
+### Bugfix
+
+- Fixed `loadedChunks` logger unexpectedly displayed in chatbox